fix: schedule cq when cq is dropped and sql node offline
