+++ app/ts-meta/meta/continuous_query.go
@@ -17,36 +17,68 @@ limitations under the License.
 package meta
 
 import (
+	"container/list"
 	"fmt"
 	"sort"
 	"time"
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/hashicorp/raft"
-	mproto "github.com/openGemini/openGemini/open_src/influx/meta/proto"
+	proto2 "github.com/openGemini/openGemini/open_src/influx/meta/proto"
+	"go.uber.org/zap"
 )
 
-func (s *Store) NotifyCQLeaseChanged() error {
-	val := &mproto.UpdateEventCommand{}
-	t := mproto.Command_UpdateEventCommand
-	cmd := &mproto.Command{Type: &t}
-	if err := proto.SetExtension(cmd, mproto.E_UpdateEventCommand_Command, val); err != nil {
-		panic(err)
-	}
-	return s.ApplyCmd(cmd)
+type situation int
+
+const (
+	sqlJoin situation = iota
+	sqlOffline
+	cqCreated
+	cqDropped
+)
+
+var (
+	// heartbeatToleranceDuration represents ts-sql heartbeat tolerance duration for crash
+	heartbeatToleranceDuration = 5 * time.Second
+
+	detectSQLNodeOfflineInterval = time.Second
+)
+
+// cqLeaseInfo represents metadata about a ts-sql.
+type cqLeaseInfo struct {
+	LastHeartbeat *list.Element
+	CQNames       []string
 }
 
-func (s *Store) applySql2MetaHeartbeat(host string) error {
+type HeartbeatInfo struct {
+	Host              string
+	LastHeartbeatTime time.Time
+}
+
+// getContinuousQueryLease return continuous query lease by specify sql host
+func (s *Store) getContinuousQueryLease(host string) ([]string, error) {
+	if !s.IsLeader() {
+		return nil, raft.ErrNotLeader
+	}
+
+	s.cqLock.RLock()
+	defer s.cqLock.RUnlock()
+
+	leaseInfo, ok := s.cqLease[host]
+	if !ok {
+		return nil, nil
+	}
+
+	return leaseInfo.CQNames, nil
+}
+
+func (s *Store) handlerSql2MetaHeartbeat(host string) error {
 	if !s.IsLeader() {
 		return raft.ErrNotLeader
 	}
-	s.updateSqlNode(host)
-	return nil
-}
-
-func (s *Store) updateSqlNode(host string) {
 	s.cqLock.Lock()
 	defer s.cqLock.Unlock()
+
 	element := s.HeartbeatInfoList.PushBack(&HeartbeatInfo{
 		Host:              host,
 		LastHeartbeatTime: time.Now(),
@@ -56,7 +88,7 @@ func (s *Store) updateSqlNode(host string) {
 	if s.cqLease[host] != nil {
 		s.HeartbeatInfoList.Remove(s.cqLease[host].LastHeartbeat)
 		s.cqLease[host].LastHeartbeat = element
-		return
+		return nil
 	}
 	// new sql node
 	s.cqLease[host] = &cqLeaseInfo{
@@ -64,9 +96,138 @@ func (s *Store) updateSqlNode(host string) {
 	}
 	s.sqlHosts = append(s.sqlHosts, host)
 	sort.Strings(s.sqlHosts)
+
 	s.scheduleCQsWithoutLock(sqlJoin)
+
+	return s.notifyCQLeaseChanged()
 }
 
+func (s *Store) notifyCQLeaseChanged() error {
+	typ := proto2.Command_NotifyCQLeaseChangedCommand
+	desc := proto2.E_NotifyCQLeaseChangedCommand_Command
+	value := &proto2.NotifyCQLeaseChangedCommand{}
+	cmd := &proto2.Command{Type: &typ}
+	if err := proto.SetExtension(cmd, desc, value); err != nil {
+		return err
+	}
+	body, err := proto.Marshal(cmd)
+	if err != nil {
+		return err
+	}
+	return s.apply(body)
+}
+
+// detectSqlNodeOffline is a background goroutine task to detect
+// which sql host has been offline.
+func (s *Store) detectSqlNodeOffline() {
+	defer s.wg.Done()
+	ticker := time.NewTicker(detectSQLNodeOfflineInterval)
+	for {
+		select {
+		case <-s.closing:
+			return
+		case <-ticker.C:
+			s.checkSQLNodesHeartbeat()
+			ticker.Reset(detectSQLNodeOfflineInterval)
+		}
+	}
+}
+
+// checkSQLNodesHeartbeat detects all sql nodes whether offline or not.
+func (s *Store) checkSQLNodesHeartbeat() {
+	var hbi *list.Element
+	if hbi = s.HeartbeatInfoList.Front(); hbi == nil {
+		return
+	}
+
+	for {
+		if time.Since(hbi.Value.(*HeartbeatInfo).LastHeartbeatTime) >= heartbeatToleranceDuration {
+			sqlHost := hbi.Value.(*HeartbeatInfo).Host
+			s.handlerSQLNodeOffline(sqlHost, hbi)
+			if hbi = s.HeartbeatInfoList.Front(); hbi == nil {
+				return
+			}
+			continue
+		}
+		break
+	}
+}
+
+// handlerSQLNodeOffline removes sql host when one ts-sql is down.
+// delete sqlHost from cqLease and sqlHosts and HeartbeatInfoList
+func (s *Store) handlerSQLNodeOffline(sqlHost string, hbi *list.Element) {
+	if !s.IsLeader() {
+		return
+	}
+
+	s.cqLock.Lock()
+	defer s.cqLock.Unlock()
+
+	// delete from cqLease
+	delete(s.cqLease, sqlHost)
+
+	// delete from sqlHosts
+	sort.Strings(s.sqlHosts)
+	n := sort.SearchStrings(s.sqlHosts, sqlHost)
+	if n < len(s.sqlHosts) {
+		s.sqlHosts = append(s.sqlHosts[:n], s.sqlHosts[n+1:]...)
+	}
+	// delete from HeartbeatInfo
+	s.HeartbeatInfoList.Remove(hbi)
+
+	s.scheduleCQsWithoutLock(sqlOffline)
+
+	// notify all alive sql nodes that CQ may be changed
+	err := s.notifyCQLeaseChanged()
+	if err != nil {
+		s.Logger.Warn("notify cq lease command failed", zap.String("sql host", sqlHost))
+	}
+}
+
+// handleCQCreated sorts s.cqNames and inserts the new cq to the s.cqNames slice.
+// At the same time assign sql node to the new cq.
+func (s *Store) handleCQCreated(newCQ string) {
+	s.cqLock.Lock()
+	defer s.cqLock.Unlock()
+
+	sort.Strings(s.cqNames)
+	n := sort.Search(len(s.cqNames), func(i int) bool {
+		return newCQ <= s.cqNames[i]
+	})
+
+	if n >= len(s.cqNames) {
+		s.cqNames = append(s.cqNames, newCQ)
+	} else {
+		if s.cqNames[n] == newCQ {
+			return
+		}
+		s.cqNames = append(s.cqNames, "")
+		copy(s.cqNames[n+1:], s.cqNames[n:])
+		s.cqNames[n] = newCQ
+	}
+
+	s.scheduleCQsWithoutLock(cqCreated)
+}
+
+// handleCQDropped sorts s.cqNames and inserts the new cq to the s.cqNames slice.
+// At the same time assign sql node to the new cq.
+func (s *Store) handleCQDropped(dropped string) {
+	s.cqLock.Lock()
+	defer s.cqLock.Unlock()
+
+	sort.Strings(s.cqNames)
+	n := sort.Search(len(s.cqNames), func(i int) bool {
+		return dropped <= s.cqNames[i]
+	})
+
+	if n < len(s.cqNames) {
+		s.cqNames = append(s.cqNames[:n], s.cqNames[n+1:]...)
+	}
+
+	s.scheduleCQsWithoutLock(cqDropped)
+}
+
+// scheduleCQsWithoutLock schedule continuous query tasks
 func (s *Store) scheduleCQsWithoutLock(sit situation) {
 	if len(s.cqNames) == 0 || len(s.sqlHosts) == 0 {
 		return
@@ -74,15 +235,20 @@ func (s *Store) scheduleCQsWithoutLock(sit situation) {
 
 	switch sit {
 	case sqlJoin:
-		s.scheduleCQWhenSqlJoin()
+		s.scheduleCQs()
+	case sqlOffline:
+		s.scheduleCQs()
 	case cqCreated:
-		s.scheduleCQWhenSqlJoin()
+		s.scheduleCQs()
+	case cqDropped:
+		s.scheduleCQs()
 	default:
 		panic(fmt.Sprintf("not exist situation %d", sit))
 	}
 }
 
-func (s *Store) scheduleCQWhenSqlJoin() {
+// scheduleCQs reassign all continuous queries to profit sql host
+func (s *Store) scheduleCQs() {
 	for _, lease := range s.cqLease {
 		lease.CQNames = lease.CQNames[:0]
 	}
