+++ app/ts-meta/meta/handlers_process.go
@@ -382,7 +382,7 @@ func (h *GetMeasurementInfo) Process() (transport.Codec, error) {
 
 func (h *Sql2MetaHeartbeat) Process() (transport.Codec, error) {
 	rsp := &message.Sql2MetaHeartbeatResponse{}
-	err := h.store.applySql2MetaHeartbeat(h.req.Host)
+	err := h.store.handlerSql2MetaHeartbeat(h.req.Host)
 	if err != nil {
 		rsp.Err = err.Error()
 		return rsp, nil
