--- app/ts-meta/meta/sql_node_info_test.go
@@ -1,37 +1,0 @@
-/*
-Copyright 2022 Huawei Cloud Computing Technologies Co., Ltd.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
- http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package meta
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func Test_CqInfo_GetLoad(t *testing.T) {
-	cqi := &CqInfo{
-		RunningCqs: make(map[string]string),
-		AssignCqs:  make(map[string]string),
-		RevokeCqs:  make(map[string]string),
-	}
-	require.Equal(t, 0, cqi.GetLoad())
-
-	cqi.AssignCqs["cq0"] = "cq0"
-	require.Equal(t, 1, cqi.GetLoad())
-
-	cqi.RunningCqs["cq1"] = "cq1"
-	cqi.RunningCqs["cq2"] = "cq2"
-	cqi.RevokeCqs["cq1"] = "cq1"
-	require.Equal(t, 2, cqi.GetLoad())
-}
