+++ lib/metaclient/meta_client_test.go
@@ -866,52 +866,3 @@ func TestDBPTCtx_String(t *testing.T) {
 	ctx.DBPTStat = nil
 	require.Equal(t, "", ctx.String())
 }
-
-func TestClient_ShowContinuousQueries(t *testing.T) {
-	c := &Client{
-		cacheData: &meta2.Data{
-			Databases: map[string]*meta2.DatabaseInfo{
-				"test1": &meta2.DatabaseInfo{
-					Name: "test1",
-					ContinuousQueries: map[string]*meta2.ContinuousQueryInfo{
-						"cq1_1": {
-							Name:  "cq1_1",
-							Query: `CREATE CONTINUOUS QUERY "cq1_1" ON "test1" RESAMPLE EVERY 1h FOR 90m BEGIN SELECT mean("passengers") INTO "average_passengers" FROM "bus_data" GROUP BY time(30m) END`,
-						},
-						"cq1_2": {
-							Name:  "cq1_2",
-							Query: `CREATE CONTINUOUS QUERY "cq1_2" ON "test1" RESAMPLE EVERY 2h FOR 30m BEGIN SELECT max("passengers") INTO "max_passengers" FROM "bus_data" GROUP BY time(10m) END`,
-						},
-					},
-				},
-				"test2": &meta2.DatabaseInfo{
-					Name: "test2",
-					ContinuousQueries: map[string]*meta2.ContinuousQueryInfo{
-						"cq2_1": {
-							Name:  "cq2_1",
-							Query: `CREATE CONTINUOUS QUERY "cq2_1" ON "test2" RESAMPLE EVERY 1h FOR 30m BEGIN SELECT min("passengers") INTO "min_passengers" FROM "bus_data" GROUP BY time(15m) END`,
-						},
-					},
-				},
-			},
-		},
-		metaServers: []string{"127.0.0.1:8092"},
-		logger:      logger.NewLogger(errno.ModuleMetaClient).With(zap.String("service", "metaclient")),
-	}
-	rows, err := c.ShowContinuousQueries()
-	if err != nil {
-		t.Fatal("query continuous queries failed!")
-	}
-
-	require.EqualValues(t, 2, len(rows))
-	for i := 0; i < rows.Len(); i++ {
-		switch rows[i].Name {
-		case "test1":
-			require.EqualValues(t, 2, len(rows[i].Values))
-		case "test2":
-			require.EqualValues(t, 1, len(rows[i].Values))
-		default:
-			t.Fatal("unexcepted query continuous queries result!")
-		}
-	}
-}
