+++ lib/metaclient/metaclient_for_cq.go
@@ -20,11 +20,113 @@ import (
 	"time"
 
 	"github.com/gogo/protobuf/proto"
+	"github.com/influxdata/influxdb/models"
 	"github.com/openGemini/openGemini/app/ts-meta/meta/message"
+	"github.com/openGemini/openGemini/lib/errno"
+	meta2 "github.com/openGemini/openGemini/open_src/influx/meta"
 	proto2 "github.com/openGemini/openGemini/open_src/influx/meta/proto"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
+
+func (c *Client) SendSql2MetaHeartbeat(host string) error {
+	startTime := time.Now()
+	currentServer := connectedServer
+	var err error
+	for {
+		c.mu.RLock()
+		select {
+		case <-c.closing:
+			c.mu.RUnlock()
+			return errors.New("Sql2MetaHeartbeat fail")
+		default:
+
+		}
+
+		if currentServer >= len(c.metaServers) {
+			currentServer = 0
+		}
+		c.mu.RUnlock()
+		err = c.sendSql2MetaHeartbeat(currentServer, host)
+		if err == nil {
+			break
+		}
+
+		if time.Since(startTime).Seconds() > float64(len(c.metaServers))*HttpReqTimeout.Seconds() {
+			break
+		}
+		time.Sleep(errSleep)
+
+		currentServer++
+	}
+	return nil
+}
+
+func (c *Client) sendSql2MetaHeartbeat(currentServer int, host string) error {
+	callback := &Sql2MetaHeartbeatCallback{}
+	msg := message.NewMetaMessage(message.Sql2MetaHeartbeatRequestMessage, &message.Sql2MetaHeartbeatRequest{Host: host})
+	err := c.SendRPCMsg(currentServer, msg, callback)
+	if err != nil {
+		c.logger.Error("Sql2MetaHeartbeat SendRPCMsg fail", zap.Error(err))
+		return err
+	}
+	return nil
+}
+
+func (c *Client) CreateContinuousQuery(database string, spec *meta2.ContinuousQuerySpec) (*meta2.ContinuousQueryInfo, error) {
+	// Benevor TODO: Check if all durations are valid
+
+	cqi := spec.NewContinuousQueryInfoBySpec()
+	// Benevor TODO: Check if the length of the cq name is legal
+
+	cmd := &proto2.CreateContinuousQueryCommand{
+		Database:        proto.String(database),
+		ContinuousQuery: cqi.Marshal(),
+	}
+
+	if err := c.retryUntilExec(proto2.Command_CreateContinuousQueryCommand, proto2.E_CreateContinuousQueryCommand_Command, cmd); err != nil {
+		return nil, err
+	}
+
+	return c.getContinuousQuery(database, cqi.Name)
+}
+
+// getContinuousQuery returns the requested continuous query info.
+func (c *Client) getContinuousQuery(database, name string) (cqi *meta2.ContinuousQueryInfo, err error) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+
+	db := c.cacheData.Database(database)
+	if db == nil || db.MarkDeleted {
+		return nil, errno.NewError(errno.DatabaseNotFound, database)
+	}
+
+	return db.GetContinuousQuery(name)
+}
+
+func (c *Client) ShowContinuousQueries() (models.Rows, error) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.cacheData.ShowContinuousQueries()
+}
+
+func (c *Client) DropContinuousQuery(name string, database string) error {
+	cmd := &proto2.DropContinuousQueryCommand{
+		Name:     proto.String(name),
+		Database: proto.String(database),
+	}
+
+	if err := c.retryUntilExec(proto2.Command_DropContinuousQueryCommand, proto2.E_DropContinuousQueryCommand_Command, cmd); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (c *Client) GetMaxCQChangeID() uint64 {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.cacheData.MaxCQChangeID
+}
 
 func (c *Client) GetCqLease(host string) ([]string, error) {
 	startTime := time.Now()
