+++ lib/metaclient/metaclient_for_cq_test.go
@@ -1,0 +1,76 @@
+/*
+Copyright 2023 Huawei Cloud Computing Technologies Co., Ltd.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package metaclient
+
+import (
+	"testing"
+
+	"github.com/openGemini/openGemini/lib/errno"
+	"github.com/openGemini/openGemini/lib/logger"
+	"github.com/openGemini/openGemini/open_src/influx/meta"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
+)
+
+func TestClient_ShowContinuousQueries(t *testing.T) {
+	c := &Client{
+		cacheData: &meta.Data{
+			Databases: map[string]*meta.DatabaseInfo{
+				"test1": &meta.DatabaseInfo{
+					Name: "test1",
+					ContinuousQueries: map[string]*meta.ContinuousQueryInfo{
+						"cq1_1": {
+							Name:  "cq1_1",
+							Query: `CREATE CONTINUOUS QUERY "cq1_1" ON "test1" RESAMPLE EVERY 1h FOR 90m BEGIN SELECT mean("passengers") INTO "average_passengers" FROM "bus_data" GROUP BY time(30m) END`,
+						},
+						"cq1_2": {
+							Name:  "cq1_2",
+							Query: `CREATE CONTINUOUS QUERY "cq1_2" ON "test1" RESAMPLE EVERY 2h FOR 30m BEGIN SELECT max("passengers") INTO "max_passengers" FROM "bus_data" GROUP BY time(10m) END`,
+						},
+					},
+				},
+				"test2": &meta.DatabaseInfo{
+					Name: "test2",
+					ContinuousQueries: map[string]*meta.ContinuousQueryInfo{
+						"cq2_1": {
+							Name:  "cq2_1",
+							Query: `CREATE CONTINUOUS QUERY "cq2_1" ON "test2" RESAMPLE EVERY 1h FOR 30m BEGIN SELECT min("passengers") INTO "min_passengers" FROM "bus_data" GROUP BY time(15m) END`,
+						},
+					},
+				},
+			},
+		},
+		metaServers: []string{"127.0.0.1:8092"},
+		logger:      logger.NewLogger(errno.ModuleMetaClient).With(zap.String("service", "metaclient")),
+	}
+	rows, err := c.ShowContinuousQueries()
+	if err != nil {
+		t.Fatal("query continuous queries failed!")
+	}
+
+	require.EqualValues(t, 2, len(rows))
+	for i := 0; i < rows.Len(); i++ {
+		switch rows[i].Name {
+		case "test1":
+			require.EqualValues(t, 2, len(rows[i].Values))
+		case "test2":
+			require.EqualValues(t, 1, len(rows[i].Values))
+		default:
+			t.Fatal("unexcepted query continuous queries result!")
+		}
+	}
+}
